Allow to call with fewer parameters.
I've noticed that I really want a convenience parameter `license_url`. However callables are a bit magical, in that they are not normal python functions. They do support `kwargs` but unlike python, where you specify that in the function definition, here that variable automagically appears when it is used in the template snippet. Luckily this can be detected in python. Still it sucks a bit.

With this change it is now possible to use fewer variables in the callable, while still having access to all if needed.